Completamento Processor

+++ src/computation/NodeComputation.java
@@ -2,5 +2,5 @@ package computation;
 
 public interface NodeComputation {
 
-	public int binaryOp(int op1, int op2);
+	public float binaryOp(float op1, float op2);
 }
