+++ src/computation/Sub.java
@@ -3,7 +3,7 @@ package computation;
 public class Sub implements NodeComputation {
 
 	@Override
-	public int binaryOp(int op1, int op2) {
+	public float binaryOp(float op1, float op2) {
 		return op1 - op2;
 	}
 }
